Improve debugging, resolve possible out of memory condition (#54)
This improves debugging of handle load/unloads by adding instruments
to the function calls so that errors can be better traced, and adding
logs for the load/unload events.

These changes revealed an issue in unsealing of data which did not
correctly unload after it had completed.

To assist in detection of these issues, tests now loop a number of
times which will trigger context failures if objects are not unloaded.